Add logic to show and hide unchanged text
Adds methods for word based trimming of text depending
on a max length and uses them to generate collapsed text
for the unchanged parts in the diff view.

Option buttons are added to collapse/uncollapse unchanged
text with JS. These options only show with JS enabled.

The page jumps a bit when JS enables the options but the
same thing happens for the default editor options so it
is kind of consistent.

The design still needs finishing touches but that could be
done in a follow up.

Bug: T149721
Change-Id: I9741e0e21c6fe1ec8fd1003cbff3c6c0e9ffe368