Keep scroll position when expanding/collapsing
With the help of this patch the relative scroll position of the
visible changes div will be saved before expanding or collapsing
elements and applied again after the expanding/collapsing is
finished.

Due to the fact, that recalibrating the scrollbar could only be
done after expanding/collapsing animiations finished, the
animation of these were removed. Otherwise the view would jump
during that process.

With the new callibration and sync of scrollbars and content
useres will be less confused and content does not suddenly move
out of sight.

Bug: T156471
Change-Id: Ic7fd3fa6df94e1c031d6e7b2b136a657d8b6d2fa