Create help dialog for TwoColConflict
This patch introduces a help dialog for the edit conflict
view. The module was rewritten to be configurable and re-
usable for other extensions.

In further patches this could be improved to have the CSS
also more flexible and independent.

The CSS and JS in this patch were loaded apart from each
other. Loading CSS with the module did not quite work.

Bug: T156250
Change-Id: I23d16b0d6f8babfbe9983d727f162111c780e12b