Improve diff quality to match TableDiffFormatter
When the custom diff formater was introduced first. A few issues appeared
with many changes in the same paragraph. An threshold was implemented to
handle these multiple changes as if the whole paragraph was removed and
added again.

While revisiting the code now and checking the diff results on the two
column conflict page it became obvious, that this threshold is in most
cases making things worse.

With this patch the threshold is removed and resulting diffs look way
better now and match the original TableDiffFormater output.

Bug: T161784
Change-Id: I51ce7eb37ac4a695f894533166ffca04d4d2d1e7