Add selection dialog for the base version
This will add the base version selection screen in the JS version. Some
magic is needed since the no-JS version should look the same and everything
else is only visible / altered with JS.

The CSS for the window dialog was added manually since this is not set for
the non-modal version of the dialog by OOjs UI. I filed a ticket for that,
I am not sure if this is intended. See T163885

Bug: T162147
Change-Id: I841b7eed23a5850506b8b6a0e62499f9332519c4