Use <br> tags in changes div
When using white-space: pre-wrap line break chars can be used to break
lines in the content and actually see the line break when viewing the
page. But the downside of this seems to be, that when you copy text
these line breaks will not be part of your copied content.

Therefore this patch usees <br> tags now instead of line break chars
inside the changes div.

Bug: T164716
Change-Id: Ib910bc00fd3a5b42cdcda759ccefac8c393d581e