Add SpecialPage to test Two Column Edit Conflicts
This patch will add a SpecialPage, that helps to test the edit
conflict interface from the Two Column Edit Conlfict extension.

Users can select an existing page as a base version for the edit
conflict. They can edit the text of the current revision of that
page and can provoke an edit conflict with a fictitious second
editor that did random changes on the same revision.

The SpecialPage can only be reached directly via
Special:SimulateTwoColEditConflict

Including:
- avoid tracking in test mode
- avoid triggering a real save
- show only when beta feature is enabled / betamode deactivated
- i18n for the fake username and fake edit summary

Bug: T170375
Change-Id: I502b6411a3a72632a7f3d52bc3b7e1c649392949