Avoid hooking into EditPage's editFormTextTop and editFormTextBottom
Wrapping the edit form with a <div> is a bit fragile, so replace the two
classes with something more robust.

Set 'mw-twocolconflict-page' on <body> and use that as a CSS selector
for increased specificity and ensuring it only applies to pages that are
TwoColConflict ones.

Next, instead of adding a mw-twocolconflict-before-base-selection class
in PHP and removing it once the base selection has been picked, look for
the absence of mw-twocolconflict-after-base-selection class with :not()
and add the class after the base selection is picked.

Change-Id: I8b37ef6459d29b93af79333e1814a856c9d5a36a