Reject non-editable pages in conflict test page
If the content model of a page does not support direct editing (e. g. a
Structured Discussions topic page, a Wikibase item, etc.), attempting to
use it in Special:SimulateTwoColEditConflict will first expose the
internal serialized content, which the user should never need to see,
and then result in an MWException. To fix this, reject such pages.

Bug: T182668
Change-Id: I01abae3e9080a9dc51872d7975c1fc6a555edff8