Add missing @covers tags and simplify existing @covers
The @covers I'm removing are specific per method. This is a good aproach
for very big classes that do to many things. A single @covers for the
entire class would create "accidential coverage", that is coverage for
code a test is not intended to cover. But this is not an issue with the
two classes in question here. They are super small and have one very
specific purpose. Accidential coverage can not happen there.

Change-Id: Ibe57e414d1ea83180a330a6c1046ec38a2a35381