Remove onclick="_blank" from test special page
I assume this was meant to be target="_blank". But I decided to remove
it for now because as the code is now it never had any effect (except
for an error message on the JavaScript console). When we change this to
target="_blank" now, a new page would open that did not opened before.

Change-Id: I26ce96ea86f7af94e263559ad572fe2a55d8b0f7