Move setContextTitle() call to TwoColConflictTestEditPage
This call is not optional any more.

This is a direct follow up to I55fbdf6, trying to make it more obvious
how this code is meant to be used.

This patch also tries to reduce the amount of code that is copied from
\EditPage::getEditButtons(). Instead, only what is different is set now.
The result is the exact same (manually confirmed).

Change-Id: I666ce107d412c3621ef4dfbe8a7c375ee870c066