Make paragraphs editable for non-js users
This addes editability for non js users. All text becomes textareas
with fitting styles. Since flexible textareas are not possible without
js a fixed size is applied by row size for now.

Word level diff colors are not present for non-js and are removed from
the headers. - By default a newline will be added to the editor text
this is common practice for MediaWiki editing. The line will be removed
when the composing of the text.

As noted in a comment in the hooks file this might go to a seperate service.

Bug: T198832
Change-Id: Ib2b332f0e9c04951b59818cdd6745632160cdf1d