Use earlier "EditPage::importFormData" hook instead of "EditPage::attemptSave"
Besides the code being a bit cleaner (notice how the test setup became more
simple) the main advantage of this other hook is that we can now use the
[[Special:SimulateTwoColEditConflict]] page to test everything. Just
simulate a conflict, and continue clicking the buttons until you reach the
"Show changes" diff view. The result of the simulated conflict resolution
will now show up there. It didn't before (for me, or I was doing something
wrong).

Change-Id: I142fcd73ab19ace7ef9d0b1873dc27fc00c1f96a