Fix misplaced active spots in the tour
The issue is very easy to reproduce by resizing the browser window
before clicking the "i" icon that starts the tour. The misplacement
happens because all positions are calculated on construction time
not when the tour starts. This patch changes just that. The data is
stored as it comes in (this is nice because no interface needs to
change). The individual steps of the tour are only created when the "i"
icon is clicked.

This also speeds up the loading time of the entire page a little bit.

Be aware that the same bug can still happen when the tour is started
a second time, and the window size changed in the meantime. This is
a little more complicated to fix, I think.

This patch also fixes a series of minor issues in this code:
* $this is not a jQuery object, and should not be prefixed with $
  because of this.
* Others are jQuery objects, and should be prefixed with $ …
* … as well as documented as such.

Bug: T195713
Change-Id: Ibd7511786499960189a4e32a042263c2a886451b