Add basic browser test for the merge view
This is the genral preparation for the newinterface browser tests and
includes the generation of a simple edit conflict. For the time being
we have the CI operate on both interfaces. Ruby browser tests will use
the default inline merge view. The NodeJS test will use the new split
interface.

To allow both interfaces to be accessed by CI a cookie will be added
as new config parameter for the interface switch. This is a temporary
code change that will be removed eventually with the deprecation of the
inline UI and the Ruby browser tests.

Due to the extension being in BetaFeature mode, we need to operate with
real user accounts. The user viewing the conflict needs to activate the
beta feature. The secound that's part of the conlfict will be created
and will use the API to create the 2nd edit.

Bug: T203141
Change-Id: Ic5898240c2b90265b5b1a877d1a2166ffc094780