Stop forwarding from the simulation special page to the real page
I had to bisect the code to find the commit that caused the issue: In
I55fbdf6 a call to `EditPage::setContextTitle` was added. This became
necesarry because core changed, and not calling this method now causes a
log error, instead of just a warning.

The error I made was setting the title to the article that was used for
the simulated conflict. That's why the user ended on that article page.
Instead the title must stick to the simulation special page. Then it works
as intended before.

Bug: T202881
Change-Id: Ifa18294bebe98fa89100c7cea92db45900e1decc