Add abstract SpecialPageHtmlFragment and start using it
All classes with a getHtml method share much more than that: They all
depend on a SpecialPage and it's ability to create Message objects as well
as give information about the User, Language, and much more Context related
stuff. I suggest to unify this.

Note the constructor is intentionally made final.

Change-Id: Ia7e2858daa1481d49849fb7c43d3f6b66b1e894d