Narrow HtmlSplitConflictHeader to not rely on it's (only) user
This patch solves 3 intertangled problems:

1. The class HtmlSplitConflictHeader (an HTML generator) had *full* access
to the class SplitTwoColConflictHelper where it was used in, almost creating
a "circular" reference between the two. Instead, the HTML generator now only
gets some much smaller value objects it actually needs.

2. The effectively only reason a WikiPage is created is to read the latest
revision. This should be an implementation detail, but was exposed.

3. While I was touching this code anyway, I implemented the suggestion
from T203085 to make sure the latest revision is always queries from the
master database.

Bug: T203085
Change-Id: Ibfb06481f4aae9f765470271dee03ad1bc5ae33d