Move buttons in OOUI popups into the $footer
This is relevant on small screens. When the content of the popup starts
to scroll, the button is the first element that disappears. This is a
massive UX issue. We might not be able to prevent scrolling entirely
(even if we try, also see T203066). But if it happens, it should happen
in a nice way.

Bug: T205379
Change-Id: I9a57a8b87706e4431657bc10a158c3efddddb4b6