Re-enable diff and preview buttons
This re-enables the diff and preview buttons for the conflict
new resolution interface. Since the resulting text is build in
onImportFormData these thing now work right way.

Adding the conflict view underneath diff and preview like described
in the ticket is done in a follow up.

Also adds basic browser tests to check that buttons are working.

Bug: T195724
Change-Id: Id79178be44b62a487366374688afc14af0139680