Add conflict screen to preview and diff
This patch adds the TwoColConflict merge view to the EditPage preview
and diff screens when they are accessed from the conflict.

Looking at the EditPage workflow changing the internal conflict state
at the point where the form is build seems to be fairly easy way to
trigger the drawing of the conflict solving screen with all implications.

The input for the injected conflict will be filled with results from the
posted and composed form data that also builds preview and diff. The step
can then be repeated showing the preview or diff of the altered text.

Since neither the preview, diff or conflict mode touch anything that would
save content unquestioned it is assumed that the implementation wont mess
with content there.

TODO:
- Tests for the new feature in SplitConflictMerger will be added via a
  separate patch.
- Browser tests will be added in a later patches.

Bug: T195724
Change-Id: If4d484581dc8c6eb752145ddf47d3aaacac0a2e4