Remove current time from "Not published yet" message
As specified in T202053. This message is now the final one we will keep.

Note that the code still tries to put the two parameters into this
message. This is fine. There are no $1 and $2 placeholders in this message
any more. The parameters will be ignored. A later patch will change the
code to clean these remaining pieces up properly.

Bug: T202053
Change-Id: I7885224a0a4c952d746386a2e3a210553cb7c535