Increase timeout in tour browser tests to 60 seconds
The default seems to be 10 seconds. Currently all patches are blocked
because they always hit this timeout. This is obviuously not a proper fix
but a hack. I assume the issue is actual load on the (virtual) machine
that runs these tests. If this is the case, 10 seconds are indeed not
enough, and the only solutions are to either free resources on the
machine, or increase the timeout.

Change-Id: I3a78679680feab7c7e606102aad1c74038aaebc6