Use MediaWikiTestCase::getTestUser instead of User::newFromName
This is relevant for some patches for T202053, because these test users
are not only identified by name, but also by ID. Currently the ID is
not tracked. With this change the ID can be obtained via
$this->otherUser->getId().

This is split from some later, larger patches for T202053 to allow us
moving forward in a few smaller steps.

Change-Id: If1e0d045288eeb5db7046784009cd37c03977d3f