Perfect initial text editor size in JS editing mode
The idea is as follows:

On the non-editable side, there is a <div> thats intentionally designed
to use the editor font, to make the text on both sides match as good as
possible. I'm using this <div>, measure it's height, and apply this height
to the <textarea> on the editable side.

I do this both ways: I measure both <div>s, use the maximum height, and
apply it to both <textarea>s (even the invisible one).

In addition:
* min-height makes sure at least 5 lines of text are visible.
* I remove a "padding" to make the <div> and <textarea> match a little
  better.
* "resize" allows the user to change the size later, if needed. This is
  expected behavior for all text editors nowadays. It's not strictly part
  of the acceptance criteria, but mentioned in the ticket.

Bug: T207163
Change-Id: I08a427d81777d51f8c8100f1459c60fbd0693dad