Remove obsolete word-wrap alias from all CSS
According to
https://developer.mozilla.org/en-US/docs/Web/CSS/overflow-wrap
word-wrap is a non-standard alias for overflow-wrap. If I'm right, the
compatibility chart shows that all browsers we want to support are able
to understand overflow-wrap. The alias is not needed.

I'm removing this now because the patch Id364efe introduces another use
of overflow-wrap, without the alias. This should be consistent one way
or the other.

Change-Id: Ic6a7efb9a87903ad97b9b87e35c6a4279f16c660