Disable ancient "vars-on-top" ESLint rule
We started discussing this in I08a427d. I decided to split it to a
separate patch.

There is indeed a reason this rule exists: "Hoisting" in JavaScript
means that "var" declarations are magically pulled to the top of a
scope. In theory this can lead to confusing issues.

In reality I have never seen this issue actually popping up anywhere
in my 20 years of experience with JavaScript. But I see confusing, hard
to read code like the one I'm touching in this patch on a daily basis.
Functions start with "var foo, bar, baz;" for no obvious reason, even
if these vars are not going to be used for possibly dozens of lines of
code.

An (in my opinion) good argument for turning this rule off is: Our
scopes are small. We try hard to split code into self-explanatory
snippets. The chance to run into an issue with hoisting (e.g. when a
later part of the same function accidentially tries to re-declare a
variable with the same name) is effectively non-existent.

Change-Id: I43bd80a5d8110ed9e292b48291f7345e6a7cc28a