Add type hints to tests using TestingAccessWrapper
This fixes some confusing type safety warnings in the IDE. The code
consuming these objects really does not care if it is a
TestingAccessWrapper, but cares about the original type.

Yea, now the IDE warns because the tests call private methods. ;-)
That's the nature of tests using TestingAccessWrapper and the reason
why it's better to avoid such things, if possible. However, now the
code is at least honest about what's going on.

Change-Id: If496908ab2fd8f307f16a39a08ae1fd27982ed91