Two-phase <textare> height calculation when activating edit mode
The idea is as follows: Before switching to the "editing" view (this
hides the <div> with the colored text, and makes the <textarea> visible)
the height of the one side where the user clicked the edit button is
measured. The <textarea> will than have at least the same height, or
more if needed. The later heavily depends on the browsers monospace font,
and if the text consumes more or less space when switching the font. For
this reason it's still good to measure the size again *after* switching to
the other font, and use whatever consumes more space.

This change greatly reduces the "jump" effects when clicking the edit pen,
which always makes me happy. :-)

The "margin: 0" removes a browser default of 1px from the <textarea>, also
to reduce the "jumping" to a minimum.

Bug: T207163
Change-Id: I86ee67259fa5ede8abbd572c330b4f2c2bae33d9