Simplify <textarea> height calculation a lot
After discussing this with the UX team we think it's better to keep this
code as dumb as possible for the first MVP. The *only* feature left in the
code makes sure the <textarea> is exactly as high as the non-editable box
was before. The user can make the <textarea> higher. This will be respected
then and not made smaller.

Removed features:
* Don't try to do anything with the window height. The problem with this is
  that it makes a second scrollbar appear, and this is always a bit iffy and
  confusing for users.
* Do not enlarge the <textarea> based on the size of the *other* side. The
  problem with this is that the user might get an editor with an huge amount
  of empty space at the bottom.

This reverts most of what Ib66ef8b and I08a427d did. We will look into this
again (and possibly re-introduce some of the behavior) later, after we got
user feedback.

Bug: T207163
Change-Id: Ia1685c86d5e4142cf94add4be6924b45195adf80