Fix harmless edits turning into conflicts
The idea is as follows: It only makes sense to stick to our conflict
resolution view when the previous POST action was done from a conflict
resolution screen. We can detect this via multiple POST parameters. I
decided to use only one. Note the same is already used a few lines above.

SWAT! This must be backported now. :-(

Bug: T205942
Bug: T208840
Change-Id: Iffec3b1e69b575eb62dc839828af449170188429