Disable diff button on conflict solving screen
The conflict solving screen already resembles a diff view that shows
both sides and furthermore, since we originally want to show the
conflict solving screen along with a diff it's a bit unclear how
this could be done without confusing the users.

See a longer explanation in the ticket in T210501#4803505.

Bug: T210501
Change-Id: Ie98966177ef81c8aa8b2621ef7a91b43dab7d536