Remove not needed subclassing in LineBasedUnifiedDiffFormatter
No code from the base class is used (no parent::… calls, no $this->…
that refers to a protected method, no constructor).

No type hint against the base class exists.

Removing the "extends" allows to add the "Diff" type hint.

I'm also inlining all code that makes it possible to inject arbitrary
HTML attributes, as it is not needed as well. All usages use the same
class name, and it's always the same for both <del> and <ins> tags.

Bug: T212200
Change-Id: I5adfdadd03b415e0faf3262e9bb3d35eacfae2fa