Mark string containing escaped HTML as such via their var names
I feel this makes the code much more readable. it makes it much more obvious
which string is going to be used in an Html::rawElement(), and why.

Currently these strings contain HTML snippets with <del> and <ins> tags, to
be shown in a diff. This does have some consequences:

* Some HTML is generated in LineBasedUnifiedDiffFormatter, and some in
HtmlSplitConflictView.

* It's a little unpredictable and potentially dangerous to use the
intermediate diff format created by LineBasedUnifiedDiffFormatter::format(),
because some elements are plain, and some are HTML.

* See the FIXME I left in one place.

Bug: T212200
Change-Id: I6cc490ddb83c4628061b8aa2d1313bb2a5a94f0e