Remove pause() from tests
The pause was done to give the browser time to load the required
ResourceLoader module. But there is a dedicated JavaScript call for
this. I believe this is the better approach.

The one detail I did not understood so far is the "return". Why was
there a return before? It should not do anything.

Change-Id: Idac9b977bf371c9253e795cf4f8c2cee634a5ec4