Avoid expensive WebRequest::getVal() where possible
getVal() is expensive because it does Unicode normalization. This is not
needed when we know the parameter contains plain ASCII only, or when it
is only used as a boolean flag. The worst thing that can happen is that
getRawVal() returns an invalid string, but that's fine when it is
immediately discarded.

Change-Id: Ie53ca6f7592347473488f6440754f9c0bc61b480