Merge ….BaseVersionSelector module into ….initJs module
This is a direct follow up to Idfae40b. The patch accidentially left
the .css file in the wrong module. Furthermore, a separate
"….BaseVersionSelector" module is not even needed. It is used a single
time, not dynamically loaded or anything. We can as well merge it in,
reducing loading time even further.

Change-Id: Ib565797a3fde3770f64e4779f56d775d7bf1dfa7