Replace slow Hamcrest assertions in HtmlSplitConflictHeaderTest
I propose this for two reasons:

1. I don't think it's relevant which HTML element contains a message.
As long as the message is present, does it matter if it is wrapped in
a <span>, <div>, or something else? Note the assertions do not even
check for specific class or id attributes. It can really be any <span>,
anywhere on the page.

2. The Hamcrest HTML assertions are suprisingly slow.

Change-Id: I2babf6ed48a24cf288f2a85f07b00ab727db1d10