Hide the floating save-bar
The save bar is problematic, it moves around depending on the viewport
size and scroll offset, and randomly obscures elements such as the
checkbox we need to click in this test.  To stabilize the tests, we'll
hide the element while clicking, then show again so it can receive
the "Save" click.

Bug: T234311
Change-Id: I96b3d0f7eb04ef3752f9a4336a092983ac04a0b6