Added carriage returns cause auto-merging to fail
It would appear that all \n in the copy text are replaced with \r\n
at some point when the user submits their conflict resolution. So
if another edit was made while this conflict was being resolved, a
new conflict will be created when the conflict resolution is finally
submitted. This is because they both technically have different copy
texts as a result of the added \r. This patch removes any \r from the
entire wikitext after merging the conflict resolution lines.

Bug: T230551
Change-Id: I73fa5bbef6ca56b77ce86fa250597316d5d73409