Consider third conflicting edits
Edits done while solving a conflict should trigger a new conflict with
the resolved version of the initial conflict on the yours side and the
data of the conflicting third edit on the other side.

This patch removes dead code that messed with the workflow in these cases.
The removed code originally was added to allow going back to the current state
of the conflict screen when using POST requests to trigger preview or diff.
This is not done anymore so the code can completely be remove.

This patch ( at least temporarly ) selects the other side by default now.

We might add something similar again when dealing with non-JS handling of
of the conflict.

Bug: T245730
Change-Id: Iba19d8fd1c372d1697a894b0ad1ccde8b4d84d16