Replace custom fade-out background colors with opacity
This fixes two issues:

* The code was quite complex with these custom colors that are not
really doing anything but simulating opacity. All modern browsers
have support for native opacity.

* When editing, the text color is changed back to black ("initial"),
but not the background colors of the <ins> and <del> elements. This
heavily reduces the usefulnes of the other side of the diff, because
the highlighting behind the black text becomes almost invisible on
brighter screens and in brighter environments.

Note: Intentionally using 0.6, even if this is not 100% identical to
before. 0.5 is *brighter* than before and *reduces* readability. 0.6
is slightly darker and increases readability, without loosing the
"this is disabled" effect.

Change-Id: Iae66d936df36eaf32796ab6c538a993a2ff623c1