Restore broken focus indicator on radio elements
This code was introduced in I40895f0. I believe this was intentionally
done to get rid of the thin white circles that show up when the element
does have focus (relevant when using the tab key to jump around). But
these white circles do have a purpose.

Related to T245112. The issue showed up while working on that
accessibility ticket.

Bug: T245507
Change-Id: I90a41328d339d01ae6faedc15a22805b900250da