Log conflict UI details at exit
These metrics make it possible to detect some cases where data is
lost during conflict resolution, in post-processing.  We'll use these
results to decide where to write custom supporting logic.

Each time the browser unloads a page in the two-column conflict
workflow, a detailed snapshot of the UI state is sent to
EventLogging.

Answers the following questions:
* Which columns were selected, and were any left unselected?
* Was the page exited by pressing the cancel button?
* Was the page exited by pressing the save button?
* How many times was the conflict page loaded for a given conflict?

The keys are designed to be joined to Schema:TwoColConflictConflict.

Bug: T245420
Change-Id: Ic5615a3b6bf48d22c7318296ba9265ca60e18530