Fix accidental data-loss when invalid POST requests are made
There was effectively no validation in place, only a few extremely basic type
checks. The merge algorithm in SplitConflictMerger simply skipped elements
it could not fully understand. The result would *always* be stored, actually
*deleting* paragraphs!

This patch introduces separate validation.

The actual merge algorithm does *not* do any validation any more, but merges
whatever it got, aiming for a result that is as *complete* as possible, even
if indivudual elements might be bogus. These are detected by the validation.

The failsafe mechanism in this patch is based on one assumption: Core will
trigger the same conflict again, if (and only if):
* The submitted editRevId is *not* the latest revision.
* The content in EditPage::$textbox1 still triggers a conflict.
The later only happens if we are intentionally submitting the users edit
again (as if "my side" was selected).

This is vaguely related to T245505 in so far that it is typically only
relevant when our frontend code is skipped, e.g. in an attack scenario, or
when a bad proxy/ad-blocker messes with the form or POST request.

Change-Id: I2c8a403d6d2fb9fa3c481dc274312d921a31a715