Remove extra submit flag in favor of content values
The extra flag does not do anything what the mw-twocolconflict-split-content
argument could not do as well. The …-content argument contains the values
from the submitted <textarea>s. These are only present if, well, a proper
POST was done.

If the …-content array is empty, there is nothing to merge. The textbox1
value was emptied before in this edge-case situation. This patch fixes this.

This also removes the null checks for the two other arguments. This does not
make this code less robust:

1. Technically, there could be a single "copy" line in the …-content
argument. Neither …-selector nor …-linefeeds are needed then. They can be
empty, or missing, and *are* actually missing from the form then. That's a
valid conflict resolution. No need to throw that away.

2. Sure, the …-selector should be validated. But this is not the job of
*this* code, and never done properly anyway. See I2c8a403 for a proper
validation.

Bug: T245505
Change-Id: I63c1921eb15c66776536f48a5d0a60332aa1073d