Detect whether an edit came from VisualEditor
I'm guessing blindly, but it seems that the `veswitched` parameter
will always be set to true when an edit comes through VisualEditor.
That gives us the ability to detect Javascript for any of the normal
edit workflows.

There's still a risk that we record false negatives due to gadgets,
custom workflows, etc.

Bug: T245722
Change-Id: I517405b620283d4554970d4dbe1752642a29a6f8