Shorten EventLogging payload
TwoColConflictConflict events included the full wikitext and summary
of a conflict, but these were often (20/day) truncated due to browser
limitations and EventLogging configuration.

This patch snips everything past the first 128 characters of both
fields.  The rest of the event takes up roughly 600 bytes, so we
should be well below the allegedly common browser URL limit of 1 kiB.

Bug: T248243
Change-Id: Ifd0bc6870261027164b8b38deff44928468765e5