Add validation method for talk page special case
The extremely nasty preprocessing for invalid single-column forms was
needed because the default behavior is to merge the "other" column
into the text before recalculating the conflict, which means it would
appear duplicated and prepended to the "your" block.  These edge
cases are straining our conflict helper integration.

Adapted from I540ff3060b.

Bug: T244237
Change-Id: I4418ed86816a4526bda5bc22510057947c2314e6