Include original wikitext in diff
The diff is responsible for snipping both the word diff HTML, and the
raw wikitext for editing.  This makes it possible to drop all
alignment logic in later patches.

Before, only the pre-save transformed text was considered in the data
structure returned by the format() function. However, only the "…html"
elements in the data structure should contain the pre-save transformed
representation. The new "…text" elements must contain the plain text as
originally entered by the user.

However, this is incomplete! In at least 1 place the pre-save
transformations are missing! Not heavily critical because all this does
is creating a visual glitch, no data loss.

Change-Id: I39a0d5fae0d44ddf756276d759a09cc9f6e65d24