Add formatter <-> view integration test
This adds an integration test between the two classes. The "formatter"
creates the intermediate array structure, the "view" consumes it. All
this test effectively cares about is that no "undefined index" are
triggered by the consuming code.

I suggest to merge this before I9dd1fce, or rebase the two on top of
each other.

Change-Id: I387c2cfd70708c11d43d3ac2fcbe13bfe41ac205