Basic test coverage for all hook handlers
Main motivation for this is the patch I4418ed8, where code is added to a
hook handler, but not tested. This here is meant to be a baseline.
Specific tests are added in later patches.

I intentionally made these test not to strict. For example, all I care
about is *that* a user preference exists (to make sure the relevant code
is executed). But I don't care what the value is.

The User type hint is intentionally removed. This parameter is unused.
To many strict type hints on hook handlers do not make the code more
safe, but more fragile for changes done outside of this codebase.

Change-Id: I59e08ff283fc41b6a2096c6dde58784b56f75ac0