Protect against null fields
User input, in this case a missing request value for "wpEdittime" or
"wpStarttime", can cause EventLogging validation to crash or logspam
the page.  The schema should be fixed to make these fields optional,
but in the meantime we can protect the application by converting null
to empty strings which pass validation.

Change-Id: I51045c8da515c2c7deef260f66ee2cf331936d90