Configurable EditStepAttempt oversampling for conflicts
Normally these events are sampled at 6.25%, but we might want to
(pending a decision) record everything to understand usage.

Unfortunately, this setting isn't sticky nor does it apply
retroactively, so we'll only get data from within the conflict
workflow and not the entire edit session.

Oversampling should only be enabled temporarily, so it's defaulting
to disabled here and we'll turn on with an explicit Wikimedia site
config flag.

Bug: T249616
Change-Id: I8e6a17738c0bd461b167789da9b436b316dba8de