Make browser tests cover HTML escaping better
I checked and this does indeed fail without the patch in I20405ba being
applied. The assertion is a little strange because it does not compare
HTML but only the text that is visible in the browser, even ignoring
newlines. With double escaping that text would be "Line<span>1</span> …"
with the HTML elements being visible as plain text. With correct escaping
the text "Line1" does not change, because the <span> do not have a
visible effect.

Bug: T249986
Depends-On: I20405ba7d6535f2330ce3c1cbf00c743adbe6e14
Change-Id: I2e480c612274ab6ea06d7d8158f3cd828b1eb70c