Fix merger trimming leading newlines
The first text block is special in so far that it might contain no visible
content, but extra newlines. These should not be ignored. Otherwise a
diff with leading newlines will be trimmed, even if no edit was made.

I believe it's not possible to entirely avoid empty textareas. If the user
added or removed newlines, that's what the diff needs to show. However, we
might improve the way this is presented.

This quick fix is needed because this edge case can cause actual data loss
(even if it's just a newline that gets lost).

I changed the JavaScript merger accordingly (this is only used for preview).

I had to update the single-column special case a bit because it was now
inserting unwanted newlines in case the first row is invalid.

Bug: T250262
Change-Id: Ia190d1a1bd63e51f4089326c679319b211d18170