Fix edit warning popping up because of irrelevant newlines
Newlines at the end of each <textarea> are stripped, and replaced with
the "extra newlines" tracked in hidden fields. No matter how many newlines
the user adds or removes, this will not cause any change and should not
count as an edit.

This doesn't have a special browser test. I would argue that the tests for
the two regular cases (when no edit was made, and when a *visible* edit
was made) are enough to sufficently cover this code.

Bug: T250262
Change-Id: I78fe8389567322b3eecaf137aeb5ea691cc8ff6d