Fix insufficient newline handling in all diff related code
What this patch does, in detail:

* Remove the /\n(?!\n)/ pattern that was meant to split the text into
non-empty paragraphs, instead of lines that might be empty. This was done
for two reasons:

1. To help the DiffEngine create more meaningful diffs. As it turns out
this is not only pointless, it also causes unwanted side effects. With
newlines being merged into the previous paragraph (instead of becoming
empty lines), we got results like "a\n\n" vs. "a\n" where the first line
should be detected as being identical, but isn't because of the different
newlines. This made the talk page special case detection fail.

2. It was also done to never have empty blocks in a diff. With mild
success. Changes to newlines resulted in a "change" block with no visible
difference. The different newlines are tracked in hidden fields, but the
user can't see this. Removing the optimization does not fix this confusion.
There is still a "change" block with no visible difference, but it's empty
now. The previous, unchanged paragraph is correctly shown as an unchanged
"copy" block.

* Use the Html::textarea() method which contains special case handling in
case the text starts with a newline.

* Add documentation for said special case.

* Do not add a newline at the end of empty <textarea>. This replicates
what EditPage does. This is a minor, not strictly related bugfix.

* Update all related tests.

* Make some tests print more meaningful debug output.

Bug: T248668
Change-Id: I8c724f6f4bd68df06fad23f9431c0b08f77841fd